requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
blobfile==2.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
filelock==3.13.4
fire==0.6.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain-community==0.0.34
langchain-core==0.1.46
langsmith==0.1.50
lxml==4.9.4
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
ollama==0.1.8
orjson==3.10.1
packaging==23.2
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
termcolor==2.4.0
torch==2.3.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
